refactor: 重构进程图标和软件相关接口
