+++ src/main/ipc/app.ts
@@ -1,0 +1,32 @@
+import { app, ipcMain, shell } from 'electron';
+
+export function registerAppIpcHandlers(): void {
+  ipcMain.handle('app:get-file-icon', async (_event, filePath: string) => {
+    try {
+      let iconPath = filePath;
+      if (process.platform === 'win32' && filePath.toLowerCase().endsWith('.lnk')) {
+        try {
+          const result = shell.readShortcutLink(filePath);
+          if (result.target) iconPath = result.target;
+        } catch {
+          // ignore
+        }
+      }
+      const icon = await app.getFileIcon(iconPath, { size: 'large' });
+      return icon.toPNG().toString('base64');
+    } catch (err) {
+      console.error('[App] get-file-icon error:', err);
+      return null;
+    }
+  });
+
+  ipcMain.handle('app:open-file', async (_event, filePath: string) => {
+    try {
+      await shell.openPath(filePath);
+      return true;
+    } catch (err) {
+      console.error('[App] open-file error:', err);
+      return false;
+    }
+  });
+}
